requirements.txt
streamlit
google-cloud-bigquery
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2